Algorithms can now be switched in-script. Still trying to figure out how to switch algorithms without opening the script.
I have no problems with maze generation. Maze is generated on play.
