+++ Assets/Scripts/GenericSingletonClass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7e7c7f7f60f33345a85d38271718153
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
