+++ Assets/Scripts/MazeAlg.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MazeAlg {
+
+    public virtual void MakeMaze(ref TileType[,] Maze)
+    {
+
+    }
+
+}
