+++ Assets/Scripts/MazeGenerator.cs
@@ -3,6 +3,8 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 
+public enum TileType { path, wall };
+
 public class MazeGenerator : GenericSingletonClass<MazeGenerator>
 {
     [SerializeField]
@@ -11,7 +13,8 @@ public class MazeGenerator : GenericSingletonClass<MazeGenerator>
     [SerializeField]
     private int height;
 
-    private enum TileType { path, wall };
+    private MazeAlg mazeAlgorithm;
+
     public Transform brick;
     private TileType[,] Maze;
 
@@ -19,29 +22,28 @@ public class MazeGenerator : GenericSingletonClass<MazeGenerator>
     private List<Vector3> pathMazes = new List<Vector3>();
     private Stack<Vector2> tileStack = new Stack<Vector2>();
 
-    private System.Random rnd = new System.Random();
 
-    private Vector2 currentTile;
-    public Vector2 CurrentTile
-    {
-        get { return currentTile; }
+    //private Vector2 currentTile;
+    //public Vector2 CurrentTile
+    //{
+    //    get { return currentTile; }
 
-        private set
-        {
-            if (value.x < 1 || value.x >= this.width - 1 || value.y < 1 || value.y >= this.height - 1)
-            {
-                throw new ArgumentException("CurrentTile must be within the one tile border all around the maze");
-            }
-            else if (value.x % 2 == 1 || value.y % 2 == 1)
-            {
-                currentTile = value;
-            }
-            else
-            {
-                throw new ArgumentException("The current square must not be both on an even X-axis and an even Y-axis, to ensure we can get walls around all tunnels");
-            }
-        }
-    }
+    //    private set
+    //    {
+    //        if (value.x < 1 || value.x >= this.width - 1 || value.y < 1 || value.y >= this.height - 1)
+    //        {
+    //            throw new ArgumentException("CurrentTile must be within the one tile border all around the maze");
+    //        }
+    //        else if (value.x % 2 == 1 || value.y % 2 == 1)
+    //        {
+    //            currentTile = value;
+    //        }
+    //        else
+    //        {
+    //            throw new ArgumentException("The current square must not be both on an even X-axis and an even Y-axis, to ensure we can get walls around all tunnels");
+    //        }
+    //    }
+    //}
 
     public override void Awake()
     {
@@ -60,9 +62,10 @@ public class MazeGenerator : GenericSingletonClass<MazeGenerator>
             }
         }
 
-        CurrentTile = Vector2.one;
-        tileStack.Push(CurrentTile);
-        RandomMazeGenerator();
+        //CurrentTile = Vector2.one;
+        //tileStack.Push(CurrentTile);
+        mazeAlgorithm = new RandomizedPrims(width, height, ref tileStack);
+        mazeAlgorithm.MakeMaze(ref Maze);
 
         for (int i = 0; i <= Maze.GetUpperBound(0); i++)
         {
@@ -86,78 +89,78 @@ public class MazeGenerator : GenericSingletonClass<MazeGenerator>
     // based on Prim's Algorithm
     // does not use weights to inform generation
     // instead randomly chooses neighbors
-    private void RandomMazeGenerator()
-    {
-        List<Vector2> neighbors;
-        List<Vector2> offsets = new List<Vector2>
-        {
-            new Vector2(0, 1),
-            new Vector2(0, -1),
-            new Vector2(1, 0),
-            new Vector2(-1, 0)
-        };
-        while (tileStack.Count > 0)
-        {
-            Maze[(int)CurrentTile.x, (int)CurrentTile.y] = TileType.path;
+    //private void RandomMazeGenerator()
+    //{
+    //    List<Vector2> neighbors;
+    //    List<Vector2> offsets = new List<Vector2>
+    //    {
+    //        new Vector2(0, 1),
+    //        new Vector2(0, -1),
+    //        new Vector2(1, 0),
+    //        new Vector2(-1, 0)
+    //    };
+    //    while (tileStack.Count > 0)
+    //    {
+    //        Maze[(int)CurrentTile.x, (int)CurrentTile.y] = TileType.path;
 
-            neighbors = GetValidNeighbors(CurrentTile, offsets);
+    //        neighbors = GetValidNeighbors(CurrentTile, offsets);
 
-            if (neighbors.Count > 0)
-            {
-                tileStack.Push(CurrentTile);
-                CurrentTile = neighbors[rnd.Next(neighbors.Count)];
-            }
-            else
-            {
-                CurrentTile = tileStack.Pop();
-            }
-        }
-    }
+    //        if (neighbors.Count > 0)
+    //        {
+    //            tileStack.Push(CurrentTile);
+    //            CurrentTile = neighbors[rnd.Next(neighbors.Count)];
+    //        }
+    //        else
+    //        {
+    //            CurrentTile = tileStack.Pop();
+    //        }
+    //    }
+    //}
 
-    /// <summary>
-    /// Get all the prospective neighboring tiles
-    /// </summary>
-    /// <param name="centerTile">The tile to test</param>
-    /// <returns>Any and all valid neighbors</returns>
-    private List<Vector2> GetValidNeighbors(Vector2 centerTile, List<Vector2> offsets)
-    {
-        List<Vector2> validNeighbors = new List<Vector2>();
+    ///// <summary>
+    ///// Get all the prospective neighboring tiles
+    ///// </summary>
+    ///// <param name="centerTile">The tile to test</param>
+    ///// <returns>Any and all valid neighbors</returns>
+    //private List<Vector2> GetValidNeighbors(Vector2 centerTile, List<Vector2> offsets)
+    //{
+    //    List<Vector2> validNeighbors = new List<Vector2>();
 
-        foreach (Vector2 offset in offsets)
-        {
-            Vector2 toCheck = centerTile + offset;
+    //    foreach (Vector2 offset in offsets)
+    //    {
+    //        Vector2 toCheck = centerTile + offset;
 
-            if (toCheck.x % 2 == 1 || toCheck.y % 2 == 1)
-            {
-                if (Maze[(int)toCheck.x, (int)toCheck.y] == TileType.wall && HasThreeWallsIntact(toCheck, offsets))
-                {
-                    validNeighbors.Add(toCheck);
-                }
-            }
-        }
+    //        if (toCheck.x % 2 == 1 || toCheck.y % 2 == 1)
+    //        {
+    //            if (Maze[(int)toCheck.x, (int)toCheck.y] == TileType.wall && HasThreeWallsIntact(toCheck, offsets))
+    //            {
+    //                validNeighbors.Add(toCheck);
+    //            }
+    //        }
+    //    }
 
-        return validNeighbors;
-    }
+    //    return validNeighbors;
+    //}
 
-    private bool HasThreeWallsIntact(Vector2 Vector2ToCheck, List<Vector2> offsets)
-    {
-        int intactWallCounter = 0;
+    //private bool HasThreeWallsIntact(Vector2 Vector2ToCheck, List<Vector2> offsets)
+    //{
+    //    int intactWallCounter = 0;
 
-        foreach (Vector2 offset in offsets)
-        {
-            Vector2 neighborToCheck = Vector2ToCheck + offset;
+    //    foreach (Vector2 offset in offsets)
+    //    {
+    //        Vector2 neighborToCheck = Vector2ToCheck + offset;
 
-            if (IsInsideMaze(neighborToCheck) && Maze[(int)neighborToCheck.x, (int)neighborToCheck.y] == TileType.wall)
-            {
-                intactWallCounter++;
-            }
-        }
+    //        if (IsInsideMaze(neighborToCheck) && Maze[(int)neighborToCheck.x, (int)neighborToCheck.y] == TileType.wall)
+    //        {
+    //            intactWallCounter++;
+    //        }
+    //    }
 
-        return intactWallCounter == 3;
-    }
+    //    return intactWallCounter == 3;
+    //}
 
-    private bool IsInsideMaze(Vector2 p)
-    {
-        return p.x >= 0 && p.y >= 0 && p.x < width && p.y < height;
-    }
+    //private bool IsInsideMaze(Vector2 p)
+    //{
+    //    return p.x >= 0 && p.y >= 0 && p.x < width && p.y < height;
+    //}
 }
