+++ Assets/Scripts/RandomizedPrims.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 38f9811b85bf26e46ac334f6aeb9e70d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
